Power Plant Forecasting using KNN and SVR regressions
In this study, the data obtained during the 6-year period (2006-2011) when a power plant was operating at full capacity was examined and a comprehensive data set consisting of a total of 9568 lines was used. This data set is presented in Excel format, and an attempt has been made to predict the hourly net electricity production (PE) of the power plant, especially in the light of the information in the tab named SAYFA2 (Sheet2).

In this estimation process, advanced machine learning algorithms were used and it was aimed to model the energy production capacity of the power plant with the highest accuracy, especially by using Support Vector Machine Based Regression (SVR) and K-Nearest Neighbors Based Regression (KNNR) methods.

While SVR draws attention with its ability to capture complex and non-linear relationships, KNNR allows making strong predictions by evaluating the closest similarities between data. Thus, with the combination of powerful mathematical models and heuristic algorithms, reliable estimates of the power plant's production output have been obtained.